+++ testcrate/tests/size_of.rs
@@ -6,7 +6,7 @@ use std::{
 use triple_arena::{
     Link, ptr_struct,
     traits::*,
-    utils::{InternalEntry, LinkNoGen, Node, PtrNoGen},
+    utils::{InternalSlot, LinkNoGen, Node, PtrNoGen},
 };
 
 ptr_struct!(P0);
@@ -32,18 +32,18 @@ fn size_of_node() {
     assert_eq!(size_of::<Node<P0, (), ()>>(), 32);
     assert_eq!(size_of::<Link<P0, ()>>(), 32);
     assert_eq!(size_of::<LinkNoGen<P0, ()>>(), 16);
-    assert_eq!(size_of::<InternalEntry<P0, ()>>(), 16);
+    assert_eq!(size_of::<InternalSlot<P0, ()>>(), 16);
     assert_eq!(
-        size_of::<InternalEntry<P0, Link<P0, Node<P0, (), ()>>>>(),
+        size_of::<InternalSlot<P0, Link<P0, Node<P0, (), ()>>>>(),
         72
     );
 
     assert_eq!(size_of::<Node<P1, (), ()>>(), 32);
     assert_eq!(size_of::<Link<P1, ()>>(), 16);
     assert_eq!(size_of::<LinkNoGen<P1, ()>>(), 16);
-    assert_eq!(size_of::<InternalEntry<P1, ()>>(), 8);
+    assert_eq!(size_of::<InternalSlot<P1, ()>>(), 8);
     assert_eq!(
-        size_of::<InternalEntry<P1, Link<P1, Node<P1, (), ()>>>>(),
+        size_of::<InternalSlot<P1, Link<P1, Node<P1, (), ()>>>>(),
         56
     );
 }