+++ triple_arena/src/arena/impl_arena_trait.rs
@@ -4,10 +4,10 @@ use crate::{
     Arena, InvalidationResult,
     arena::{
         ArenaBacking,
-        InternalEntry::{self, *},
+        InternalSlot::{self, *},
     },
-    traits::{ArenaTrait, Ptr},
-    utils::{NonZeroInxGenericStack, PtrGen, PtrInx},
+    traits::{ArenaInsertTrait, ArenaTrait, Ptr},
+    utils::{AllocError, NonZeroInxGenericStack, PtrGen, PtrInx},
 };
 
 impl<P: Ptr, T, B: ArenaBacking> ArenaTrait<P, T> for Arena<P, T, B> {
@@ -28,10 +28,7 @@ impl<P: Ptr, T, B: ArenaBacking> ArenaTrait<P, T> for Arena<P, T, B> {
         self.m.max_capacity()
     }
 
-    fn reallocate_min_capacity(
-        &mut self,
-        min_capacity: usize,
-    ) -> Result<(), crate::utils::AllocError> {
+    fn reallocate_min_capacity(&mut self, min_capacity: usize) -> Result<(), AllocError> {
         // so that capacity on the end is not used up by unallocated slots, and just fix
         // up the freelist if this function was called under any circumstance, it is
         // understood that it is a `O(n)` operation anyway.
@@ -49,57 +46,6 @@ impl<P: Ptr, T, B: ArenaBacking> ArenaTrait<P, T> for Arena<P, T, B> {
 
     fn set_generation(&mut self, new_gen: P::Gen) {
         self.generation = new_gen;
-    }
-
-    fn insert_within_capacity(&mut self, t: T) -> Result<(P, &mut T), T> {
-        let generation = self.generation;
-        if let Some(inx) = self.freelist_root {
-            let entry = self.m.get_mut(Self::into_checked(inx)).unwrap();
-            let InternalEntry::Free(next) = mem::replace(entry, Allocated(generation, t)) else {
-                unreachable!()
-            };
-            if next == inx {
-                // end of freelist
-                self.freelist_root = None;
-            } else {
-                // move to next node in the freelist
-                self.freelist_root = Some(next);
-            }
-            self.len = self.len.wrapping_add(1);
-            let InternalEntry::Allocated(_, t) = entry else {
-                unreachable!()
-            };
-            Ok((Ptr::_from_raw(inx, generation), t))
-        } else {
-            // see if capacity for entries remains, freelist remains unset if we push just
-            // one thing
-            match self
-                .m
-                .push_within_capacity(InternalEntry::Allocated(self.generation, t))
-            {
-                // TODO Polonius cleans this up
-                Ok((raw_inx, _)) => {
-                    if let Some(inx) = P::Inx::try_from_usize(raw_inx) {
-                        let Some(InternalEntry::Allocated(_, t)) = self.m.get_mut(raw_inx) else {
-                            unreachable!()
-                        };
-                        Ok((<P as Ptr>::_from_raw(inx, generation), t))
-                    } else {
-                        // undo
-                        let InternalEntry::Allocated(_, t) = self.m.pop().unwrap() else {
-                            unreachable!()
-                        };
-                        Err(t)
-                    }
-                }
-                Err(entry) => {
-                    let InternalEntry::Allocated(_, t) = entry else {
-                        unreachable!()
-                    };
-                    Err(t)
-                }
-            }
-        }
     }
 
     fn get_inx(&self, p: <P as Ptr>::Inx) -> Option<(<P as Ptr>::Gen, &T)> {
@@ -122,13 +68,13 @@ impl<P: Ptr, T, B: ArenaBacking> ArenaTrait<P, T> for Arena<P, T, B> {
         let indices = indices.map(|inx| PtrInx::try_into_usize(inx).unwrap());
         match self.m.get_disjoint_mut(indices) {
             Ok(a) => {
-                for entry in &a {
-                    if matches!(entry, Free(_)) {
+                for slot in &a {
+                    if matches!(slot, Free(_)) {
                         return Err(GetDisjointMutError::IndexOutOfBounds);
                     }
                 }
-                Ok(a.map(|entry| {
-                    let Allocated(generation, t) = entry else {
+                Ok(a.map(|slot| {
+                    let Allocated(generation, t) = slot else {
                         unreachable!()
                     };
                     (*generation, t)
@@ -165,10 +111,72 @@ impl<P: Ptr, T, B: ArenaBacking> ArenaTrait<P, T> for Arena<P, T, B> {
     }
 
     fn remove(&mut self, p: P) -> InvalidationResult<T> {
-        self.remove_internal(p, true)
+        self.remove_internal(p.inx(), Some(p.generation()), true)
     }
 
     fn clear(&mut self) {
         self.m.clear();
+    }
+
+    fn clone_from_with<U, A: ArenaTrait<P, U>, F: FnMut(P, &U) -> T>(
+        &mut self,
+        source: &A,
+        map: F,
+    ) -> Result<(), AllocError> {
+        //if self.capacity() < sour
+        todo!()
+    }
+}
+
+impl<P: Ptr, T, B: ArenaBacking> ArenaInsertTrait<P, T> for Arena<P, T, B> {
+    fn insert_within_capacity(&mut self, t: T) -> Result<(P, &mut T), T> {
+        let generation = self.generation;
+        if let Some(inx) = self.freelist_root {
+            let slot = self.m.get_mut(Self::into_checked(inx)).unwrap();
+            let InternalSlot::Free(next) = mem::replace(slot, Allocated(generation, t)) else {
+                unreachable!()
+            };
+            if next == inx {
+                // end of freelist
+                self.freelist_root = None;
+            } else {
+                // move to next node in the freelist
+                self.freelist_root = Some(next);
+            }
+            self.len = self.len.wrapping_add(1);
+            let InternalSlot::Allocated(_, t) = slot else {
+                unreachable!()
+            };
+            Ok((Ptr::_from_raw(inx, generation), t))
+        } else {
+            // see if capacity for slots remains, freelist remains unset if we push just
+            // one thing
+            match self
+                .m
+                .push_within_capacity(InternalSlot::Allocated(self.generation, t))
+            {
+                // TODO Polonius cleans this up
+                Ok((raw_inx, _)) => {
+                    if let Some(inx) = P::Inx::try_from_usize(raw_inx) {
+                        let Some(InternalSlot::Allocated(_, t)) = self.m.get_mut(raw_inx) else {
+                            unreachable!()
+                        };
+                        Ok((<P as Ptr>::_from_raw(inx, generation), t))
+                    } else {
+                        // undo
+                        let InternalSlot::Allocated(_, t) = self.m.pop().unwrap() else {
+                            unreachable!()
+                        };
+                        Err(t)
+                    }
+                }
+                Err(slot) => {
+                    let InternalSlot::Allocated(_, t) = slot else {
+                        unreachable!()
+                    };
+                    Err(t)
+                }
+            }
+        }
     }
 }